Staff demonstrator request kinda works

+++ database/migrations/2017_07_21_065914_create_demonstrator_requests_table.php
@@ -18,8 +18,8 @@ class CreateDemonstratorRequestsTable extends Migration
             $table->unsignedInteger('course_id');
             $table->unsignedInteger('staff_id');
             $table->string('type')->nullable();
-            $table->unsignedInteger('hours_needed');
-            $table->unsignedInteger('demonstrators_needed');
+            $table->float('hours_needed');
+            $table->float('demonstrators_needed');
             $table->boolean('semester_1')->default(false);
             $table->boolean('semester_2')->default(false);
             $table->boolean('semester_3')->default(false);
